# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EdgeCast – Kalshi weather market monitoring API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication
  - name: dashboard
    description: Dashboard summary
  - name: markets
    description: Kalshi weather markets
  - name: weather
    description: Weather forecasts
  - name: jobs
    description: Data collection jobs
  - name: errors
    description: Application errors
  - name: paper-trades
    description: Paper trading simulation

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /health/services:
    get:
      operationId: getServiceHealth
      tags: [health]
      summary: Status of each external API, database, and scheduler
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Service health detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceHealthResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: JWT token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthToken"
        "401":
          description: Invalid credentials

  /dashboard:
    get:
      operationId: getDashboard
      tags: [dashboard]
      summary: Dashboard summary with last-run stats
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dashboard data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dashboard"

  /markets:
    get:
      operationId: getMarkets
      tags: [markets]
      summary: All collected Kalshi weather markets
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Market list with optional summary message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketListResponse"

  /markets/{ticker}:
    get:
      operationId: getMarket
      tags: [markets]
      summary: Get single market by ticker
      security:
        - bearerAuth: []
      parameters:
        - name: ticker
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Market detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Market"
        "404":
          description: Not found

  /weather/forecasts:
    get:
      operationId: getWeatherForecasts
      tags: [weather]
      summary: Stored weather forecast snapshots
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Forecast list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WeatherForecast"

  /jobs:
    get:
      operationId: getJobs
      tags: [jobs]
      summary: List recent job runs
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Job run list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JobRun"

  /jobs/collect:
    post:
      operationId: triggerCollection
      tags: [jobs]
      summary: Trigger manual data collection now
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Job started or already running
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobRun"

  /errors:
    get:
      operationId: getErrors
      tags: [errors]
      summary: Recent application errors
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: number
            default: 50
      responses:
        "200":
          description: Error list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AppError"

  /paper-trades:
    get:
      operationId: listPaperTrades
      tags: [paper-trades]
      summary: List paper trades with optional filters
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: direction
          in: query
          schema:
            type: string
        - name: confidence
          in: query
          schema:
            type: string
        - name: city
          in: query
          schema:
            type: string
        - name: contract_type
          in: query
          schema:
            type: string
        - name: date_from
          in: query
          schema:
            type: string
        - name: date_to
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: number
            default: 200
      responses:
        "200":
          description: Trade list
          content:
            application/json:
              schema:
                type: object
                required: [trades, total]
                properties:
                  trades:
                    type: array
                    items:
                      $ref: "#/components/schemas/PaperTrade"
                  total:
                    type: number

  /paper-trades/metrics:
    get:
      operationId: getPaperTradeMetrics
      tags: [paper-trades]
      summary: Aggregated paper trading summary statistics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Metrics summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaperTradeSummary"

  /paper-trades/settings:
    get:
      operationId: getPaperTradeSettings
      tags: [paper-trades]
      summary: Get paper trading settings
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaperTradeSettings"
    put:
      operationId: updatePaperTradeSettings
      tags: [paper-trades]
      summary: Update paper trading settings
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaperTradeSettings"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaperTradeSettings"

  /paper-trades/{tradeId}:
    get:
      operationId: getPaperTrade
      tags: [paper-trades]
      summary: Full detail for a single paper trade
      security:
        - bearerAuth: []
      parameters:
        - name: tradeId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Trade detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaperTrade"
        "404":
          description: Not found

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer

  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ServiceStatus:
      type: object
      required: [name, status, message, lastChecked]
      properties:
        name:
          type: string
        status:
          type: string
          enum: [ok, error, unknown]
        message:
          type: ["string", "null"]
        lastChecked:
          type: ["string", "null"]

    ServiceHealthResponse:
      type: object
      required: [services]
      properties:
        services:
          type: array
          items:
            $ref: "#/components/schemas/ServiceStatus"
        lastSuccessfulCollection:
          type: ["string", "null"]
        lastFailedCollection:
          type: ["string", "null"]
        totalMarketsStored:
          type: number
        totalForecastsStored:
          type: number

    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string

    AuthToken:
      type: object
      required: [accessToken, tokenType]
      properties:
        accessToken:
          type: string
        tokenType:
          type: string

    Market:
      type: object
      required: [id, ticker, title, status, weatherMatched]
      properties:
        id:
          type: number
        ticker:
          type: string
        eventTicker:
          type: ["string", "null"]
        title:
          type: string
        subtitle:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        targetDate:
          type: ["string", "null"]
        openTime:
          type: ["string", "null"]
        closeTime:
          type: ["string", "null"]
        status:
          type: string
        yesBid:
          type: ["number", "null"]
        yesAsk:
          type: ["number", "null"]
        noBid:
          type: ["number", "null"]
        noAsk:
          type: ["number", "null"]
        volume:
          type: ["number", "null"]
        weatherMatched:
          type: boolean
        parsingStatus:
          type: ["string", "null"]
        parsingReason:
          type: ["string", "null"]
        weatherMarketType:
          type: ["string", "null"]
        collectionTimestamp:
          type: ["string", "null"]
        lastUpdated:
          type: ["string", "null"]
        priceSource:
          type: ["string", "null"]
        # Phase 2A analysis fields (populated after a collection run)
        ecProbability:
          type: ["number", "null"]
        marketProbability:
          type: ["number", "null"]
        probabilityDiff:
          type: ["number", "null"]
        confidence:
          type: ["string", "null"]
        analysisStatus:
          type: ["string", "null"]
        analysisReason:
          type: ["string", "null"]
        explanation:
          type: ["string", "null"]
        settlementVariable:
          type: ["string", "null"]
        settlementOperator:
          type: ["string", "null"]
        settlementThreshold:
          type: ["number", "null"]
        leadTimeDays:
          type: ["number", "null"]
        forecastValue:
          type: ["number", "null"]
        # Phase 2B contract type and extended fields
        contractType:
          type: ["string", "null"]
        targetHour:
          type: ["number", "null"]
        targetTimezoneStr:
          type: ["string", "null"]
        lowerBound:
          type: ["number", "null"]
        upperBound:
          type: ["number", "null"]

    MarketListResponse:
      type: object
      required: [markets]
      properties:
        markets:
          type: array
          items:
            $ref: "#/components/schemas/Market"
        summary:
          type: ["string", "null"]

    WeatherForecast:
      type: object
      required: [id, city, forecastDate, retrievedAt]
      properties:
        id:
          type: number
        city:
          type: string
        forecastDate:
          type: string
        temperatureHigh:
          type: ["number", "null"]
        temperatureLow:
          type: ["number", "null"]
        precipitationProb:
          type: ["number", "null"]
        windSpeed:
          type: ["number", "null"]
        retrievedAt:
          type: string

    Dashboard:
      type: object
      required: [totalActiveMarkets, marketsWithWeather, markets, recentErrors]
      properties:
        totalActiveMarkets:
          type: number
        marketsWithWeather:
          type: number
        marketsCollected:
          type: number
        marketsParseFailures:
          type: number
        lastCollectionTime:
          type: ["string", "null"]
        lastCollectionStatus:
          type: ["string", "null"]
        lastCollectionDuration:
          type: ["number", "null"]
        lastCollectionMarketsFound:
          type: ["number", "null"]
        lastCollectionMarketsSkipped:
          type: ["number", "null"]
        collectionSummary:
          type: ["string", "null"]
        markets:
          type: array
          items:
            $ref: "#/components/schemas/Market"
        recentErrors:
          type: array
          items:
            $ref: "#/components/schemas/AppError"
        lastJob:
          oneOf:
            - $ref: "#/components/schemas/JobRun"
            - type: "null"

    JobRun:
      type: object
      required: [id, jobType, startedAt, status]
      properties:
        id:
          type: number
        jobType:
          type: string
        startedAt:
          type: string
        completedAt:
          type: ["string", "null"]
        status:
          type: string
          enum: [running, success, failed]
        marketsFound:
          type: ["number", "null"]
        marketsSkipped:
          type: ["number", "null"]
        marketsRejected:
          type: ["number", "null"]
        forecastsRetrieved:
          type: ["number", "null"]
        durationSeconds:
          type: ["number", "null"]
        errorMessage:
          type: ["string", "null"]
        ptCandidates:
          type: ["number", "null"]
        ptCreated:
          type: ["number", "null"]
        ptYesTrades:
          type: ["number", "null"]
        ptNoTrades:
          type: ["number", "null"]
        ptSkipped:
          type: ["number", "null"]
        ptErrors:
          type: ["number", "null"]

    AppError:
      type: object
      required: [id, errorType, message, occurredAt]
      properties:
        id:
          type: number
        errorType:
          type: string
        message:
          type: string
        context:
          type: ["string", "null"]
        occurredAt:
          type: string

    PaperTrade:
      type: object
      required: [id, marketTicker, strategyVersion, direction, stake, status]
      properties:
        id:
          type: number
        createdAt:
          type: ["string", "null"]
        marketTicker:
          type: string
        eventTicker:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        weatherVariable:
          type: ["string", "null"]
        contractType:
          type: ["string", "null"]
        targetSettlementDate:
          type: ["string", "null"]
        snapshotId:
          type: ["number", "null"]
        strategyVersion:
          type: string
        direction:
          type: string
          enum: [YES, NO]
        ecYesProbability:
          type: ["number", "null"]
        ecSideProbability:
          type: ["number", "null"]
        marketYesProbability:
          type: ["number", "null"]
        sideMarketPrice:
          type: ["number", "null"]
        priceSource:
          type: ["string", "null"]
        edgePctPoints:
          type: ["number", "null"]
        confidenceScore:
          type: ["number", "null"]
        confidenceLabel:
          type: ["string", "null"]
        stake:
          type: number
        quantity:
          type: ["number", "null"]
        status:
          type: string
          enum: [OPEN, SETTLED, VOID, ERROR]
        kalshiResult:
          type: ["string", "null"]
        outcome:
          type: ["string", "null"]
          enum: [WIN, LOSS, VOID, null]
        settlementTimestamp:
          type: ["string", "null"]
        grossPayout:
          type: ["number", "null"]
        profitLoss:
          type: ["number", "null"]
        returnPct:
          type: ["number", "null"]
        decisionExplanation:
          type: ["string", "null"]
        warnings:
          type: ["string", "null"]
        market:
          oneOf:
            - type: "null"
            - type: object
              properties:
                title: { type: ["string", "null"] }
                subtitle: { type: ["string", "null"] }
                targetDate: { type: ["string", "null"] }
                yesBid: { type: ["number", "null"] }
                yesAsk: { type: ["number", "null"] }
                noBid: { type: ["number", "null"] }
                noAsk: { type: ["number", "null"] }
                weatherMarketType: { type: ["string", "null"] }
        snapshot:
          oneOf:
            - type: "null"
            - type: object
              properties:
                id: { type: ["number", "null"] }
                createdAt: { type: ["string", "null"] }
                ecProbability: { type: ["number", "null"] }
                marketProbability: { type: ["number", "null"] }
                confidence: { type: ["string", "null"] }
                explanation: { type: ["string", "null"] }
                forecastValue: { type: ["number", "null"] }
                leadTimeDays: { type: ["number", "null"] }
                settlementVariable: { type: ["string", "null"] }
                settlementOperator: { type: ["string", "null"] }
                settlementThreshold: { type: ["number", "null"] }
                contractType: { type: ["string", "null"] }
                lowerBound: { type: ["number", "null"] }
                upperBound: { type: ["number", "null"] }

    PaperTradeSummary:
      type: object
      required: [openCount, settledCount, voidCount, totalCount, wins, losses, totalStaked, netProfitLoss, sampleSizeWarning]
      properties:
        openCount:
          type: number
        settledCount:
          type: number
        voidCount:
          type: number
        totalCount:
          type: number
        wins:
          type: number
        losses:
          type: number
        winRate:
          type: ["number", "null"]
        totalStaked:
          type: number
        netProfitLoss:
          type: number
        roi:
          type: ["number", "null"]
        avgEntryEdge:
          type: ["number", "null"]
        avgWinEdge:
          type: ["number", "null"]
        avgLossEdge:
          type: ["number", "null"]
        byDirection:
          type: array
          items:
            type: object
            properties:
              label: { type: string }
              total: { type: number }
              open: { type: number }
              settled: { type: number }
              wins: { type: number }
              losses: { type: number }
              winRate: { type: ["number", "null"] }
              netProfitLoss: { type: number }
        byConfidence:
          type: array
          items:
            type: object
            properties:
              label: { type: string }
              total: { type: number }
              open: { type: number }
              settled: { type: number }
              wins: { type: number }
              losses: { type: number }
              winRate: { type: ["number", "null"] }
              netProfitLoss: { type: number }
        byCity:
          type: array
          items:
            type: object
            properties:
              label: { type: string }
              total: { type: number }
              open: { type: number }
              settled: { type: number }
              wins: { type: number }
              losses: { type: number }
              winRate: { type: ["number", "null"] }
              netProfitLoss: { type: number }
        byContractType:
          type: array
          items:
            type: object
            properties:
              label: { type: string }
              total: { type: number }
              open: { type: number }
              settled: { type: number }
              wins: { type: number }
              losses: { type: number }
              winRate: { type: ["number", "null"] }
              netProfitLoss: { type: number }
        sampleSizeWarning:
          type: boolean
        preliminaryNote:
          type: ["string", "null"]

    PaperTradeSettings:
      type: object
      properties:
        enabled:
          type: ["boolean", "null"]
        min_edge_pct:
          type: ["number", "null"]
        min_confidence:
          type: ["string", "null"]
        stake:
          type: ["number", "null"]
        strategy_version:
          type: ["string", "null"]