# lib/api-zod/src/generated/types/jobRun.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * EdgeCast – Kalshi weather market monitoring API
 * OpenAPI spec version: 0.1.0
 */
import type { JobRunStatus } from './jobRunStatus';

export interface JobRun {
  id: number;
  jobType: string;
  startedAt: string;
  /** @nullable */
  completedAt?: string | null;
  status: JobRunStatus;
  /** @nullable */
  marketsFound?: number | null;
  /** @nullable */
  marketsSkipped?: number | null;
  /** @nullable */
  marketsRejected?: number | null;
  /** @nullable */
  forecastsRetrieved?: number | null;
  /** @nullable */
  durationSeconds?: number | null;
  /** @nullable */
  errorMessage?: string | null;
  /** @nullable */
  ptCandidates?: number | null;
  /** @nullable */
  ptCreated?: number | null;
  /** @nullable */
  ptYesTrades?: number | null;
  /** @nullable */
  ptNoTrades?: number | null;
  /** @nullable */
  ptSkipped?: number | null;
  /** @nullable */
  ptErrors?: number | null;
}