# lib/api-zod/src/generated/types/paperTradeSettings.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * EdgeCast – Kalshi weather market monitoring API
 * OpenAPI spec version: 0.1.0
 */

export interface PaperTradeSettings {
  /** @nullable */
  enabled?: boolean | null;
  /** @nullable */
  min_edge_pct?: number | null;
  /** @nullable */
  min_confidence?: string | null;
  /** @nullable */
  stake?: number | null;
  /** @nullable */
  strategy_version?: string | null;
}